Sand colors (to build) and refined crates campaign
3 sand colors added to the build menu so players can use it when they
want sand (for paths or to fix the beachs, etc...)
The campaign now is smart enough to warn the player if he does not break
the crate, teaching him how to do it so the campaign can procced. And
the dialogue now happens 24 hours after breaking it, not 24h after
placing it as before.